requirements.txt
numpy>=1.24
pillow>=10.0
tqdm>=4.66
pyyaml>=6.0
pandas>=2.0
matplotlib>=3.8
ipykernel>=6.29